# feat: lfilter with filterbanks support