--- a/test/torchaudio_unittest/functional/autograd_impl.py
+++ b/test/torchaudio_unittest/functional/autograd_impl.py
@@ -61,7 +61,7 @@ class Autograd(TestBaseMixin):
 
     def test_lfilter_filterbanks(self):
         torch.random.manual_seed(2434)
-        x = get_whitenoise(sample_rate=22050, duration=0.01, n_channels=2)
+        x = get_whitenoise(sample_rate=22050, duration=0.01, n_channels=3)
         a = torch.tensor([[0.7, 0.2, 0.6],
                           [0.8, 0.2, 0.9]])
         b = torch.tensor([[0.4, 0.2, 0.9],