resolves #1528 , relates to #1561 

This merge add the ability to apply multiple filters in `F.lfilter`. 
When the input shape is `(..., time)`, if the shape of filter coefficients is `(number_of_filters, filter_order)`, the shape of output will be `(...., number_of_filters, time)`, which is the input shape *plus* one additional dimension from the filters. 


Before feeding to the cpp backend, the python frontend first duplicate the input `number_of_filters`  times and stacking them to form a shape of `(...., number_of_filters, time)`.

The cpp backend assume the shape of `waveform` and  `*_coeffs` are `(batch, number_of_filters, time)` and `(number_of_filters, filter_order)` (before are just `(batch, time)` and `(filter_order,)`). Different filters is applied parallelly at the second dimension of `waveform`.

cc @mthrok