@yoyololicon 2021-06-22
> Overall, it looks good. The semantic around filter bank looks straightforward though documentation can be improved.
> 
> If I understand correctly, the implementation cannot handle the case where one wants to apply different filters per channel. (though I am not sure if that's valid use-case.) Is that correct?

No, it cannot apply different filters per channel. All signals are applied with same set of filters, regardless.