@yoyololicon 2021-07-02
# Benchmarks
I have managed to build the sources with cuda support using docker container. 
The benchmarks was conducted on a linux machine with i5 4790K, 32G of RAM, and a RTX 3070.
The result indicate that the runtime on CUDA is roughly doubled. 

## Before

### CPU
```
[-------------- IIR filter --------------]
                   |  forward  |  backward
1 threads: -------------------------------
      [32, 256]    |    307.2  |    808.0 
      [32, 1024]   |    614.9  |   1512.9 
      [32, 4096]   |   2036.1  |   5199.6 
      [64, 256]    |    399.5  |    992.8 
      [64, 1024]   |   1027.6  |   2417.2 
      [64, 4096]   |   7028.7  |  13454.1 
      [128, 256]   |    624.4  |   1493.7 
      [128, 1024]  |   1986.3  |   4805.7 
      [128, 4096]  |  17420.1  |  31806.2 
2 threads: -------------------------------
      [32, 256]    |    291.6  |    813.7 
      [32, 1024]   |    531.0  |   1417.3 
      [32, 4096]   |   1754.9  |   4138.1 
      [64, 256]    |    357.4  |    968.6 
      [64, 1024]   |    797.2  |   1973.9 
      [64, 4096]   |   6914.0  |  13116.0 
      [128, 256]   |    535.3  |   1398.3 
      [128, 1024]  |   1522.3  |   4060.9 
      [128, 4096]  |  17863.6  |  30650.4 
4 threads: -------------------------------
      [32, 256]    |    268.3  |    822.9 
      [32, 1024]   |    491.3  |   1395.4 
      [32, 4096]   |   1539.7  |   4215.9 
      [64, 256]    |    347.6  |    996.4 
      [64, 1024]   |    708.5  |   1801.9 
      [64, 4096]   |   7666.7  |  13461.7 
      [128, 256]   |    479.9  |   1306.3 
      [128, 1024]  |   1428.4  |   4137.1 
      [128, 4096]  |  18735.3  |  31495.1 

Times are in microseconds (us).
```

### CUDA

```
[-------------- IIR filter --------------]
                   |  forward  |  backward
1 threads: -------------------------------
      [32, 256]    |    10.7   |    19.6  
      [32, 1024]   |    41.7   |    76.5  
      [32, 4096]   |   159.8   |   297.6  
      [64, 256]    |    10.3   |    19.7  
      [64, 1024]   |    41.8   |    75.9  
      [64, 4096]   |   157.9   |   299.2  
      [128, 256]   |    10.5   |    19.8  
      [128, 1024]  |    41.3   |    75.9  
      [128, 4096]  |   164.3   |   301.0  
2 threads: -------------------------------
      [32, 256]    |    10.6   |    19.6  
      [32, 1024]   |    40.2   |    77.6  
      [32, 4096]   |   161.9   |   302.6  
      [64, 256]    |    10.7   |    19.9  
      [64, 1024]   |    40.0   |    75.6  
      [64, 4096]   |   159.8   |   299.9  
      [128, 256]   |    10.9   |    19.8  
      [128, 1024]  |    41.4   |    75.2  
      [128, 4096]  |   170.6   |   301.2  
4 threads: -------------------------------
      [32, 256]    |    10.8   |    19.6  
      [32, 1024]   |    40.3   |    76.8  
      [32, 4096]   |   160.5   |   301.7  
      [64, 256]    |    10.6   |    19.7  
      [64, 1024]   |    40.3   |    75.5  
      [64, 4096]   |   163.7   |   302.8  
      [128, 256]   |    10.7   |    19.8  
      [128, 1024]  |    42.8   |    75.9  
      [128, 4096]  |   165.5   |   299.8  

Times are in milliseconds (ms).
```

## After

### CPU
```
[-------------- IIR filter --------------]
                   |  forward  |  backward
1 threads: -------------------------------
      [32, 256]    |    322.1  |    835.5 
      [32, 1024]   |    635.2  |   1538.1 
      [32, 4096]   |   2164.4  |   5424.1 
      [64, 256]    |    410.9  |   1026.8 
      [64, 1024]   |   1010.5  |   2299.0 
      [64, 4096]   |   6981.6  |  12721.7 
      [128, 256]   |    616.4  |   1451.6 
      [128, 1024]  |   1886.5  |   4810.8 
      [128, 4096]  |  17436.0  |  31758.8 
2 threads: -------------------------------
      [32, 256]    |    305.4  |    854.4 
      [32, 1024]   |    550.3  |   1464.6 
      [32, 4096]   |   1733.3  |   4914.4 
      [64, 256]    |    370.5  |   1003.6 
      [64, 1024]   |    791.4  |   1927.2 
      [64, 4096]   |   6829.2  |  12608.5 
      [128, 256]   |    536.6  |   1383.8 
      [128, 1024]  |   1429.7  |   4541.2 
      [128, 4096]  |  17876.0  |  31664.5 
4 threads: -------------------------------
      [32, 256]    |    283.0  |    850.6 
      [32, 1024]   |    492.4  |   1446.2 
      [32, 4096]   |   1586.0  |   4792.5 
      [64, 256]    |    362.8  |    978.7 
      [64, 1024]   |    735.8  |   1842.1 
      [64, 4096]   |   7331.0  |  12645.4 
      [128, 256]   |    492.1  |   1351.4 
      [128, 1024]  |   1265.6  |   3954.7 
      [128, 4096]  |  18502.5  |  33391.6 

Times are in microseconds (us).
```

### CUDA

```
[-------------- IIR filter --------------]
                   |  forward  |  backward
1 threads: -------------------------------
      [32, 256]    |    19.6   |    37.9  
      [32, 1024]   |    77.4   |   148.7  
      [32, 4096]   |   314.3   |   596.4  
      [64, 256]    |    19.7   |    38.5  
      [64, 1024]   |    82.8   |   147.6  
      [64, 4096]   |   328.1   |   589.7  
      [128, 256]   |    20.5   |    38.0  
      [128, 1024]  |    82.3   |   148.7  
      [128, 4096]  |   323.4   |   595.7  
2 threads: -------------------------------
      [32, 256]    |    19.9   |    37.9  
      [32, 1024]   |    77.4   |   150.9  
      [32, 4096]   |   308.5   |   597.6  
      [64, 256]    |    19.7   |    38.6  
      [64, 1024]   |    82.7   |   147.6  
      [64, 4096]   |   330.5   |   590.1  
      [128, 256]   |    20.8   |    38.2  
      [128, 1024]  |    80.9   |   149.3  
      [128, 4096]  |   330.8   |   595.9  
4 threads: -------------------------------
      [32, 256]    |    19.8   |    37.9  
      [32, 1024]   |    78.2   |   150.8  
      [32, 4096]   |   309.5   |   598.0  
      [64, 256]    |    20.1   |    38.1  
      [64, 1024]   |    82.6   |   148.9  
      [64, 4096]   |   323.4   |   589.2  
      [128, 256]   |    20.9   |    38.2  
      [128, 1024]  |    81.4   |   148.7  
      [128, 4096]  |   322.0   |   596.0  

Times are in milliseconds (ms).
```