@mthrok 2021-07-02
Hi @yoyololicon 

Thanks for the benchmark.
The shapes in the table, are they the size of input batch?
I am also curious to learn the performance difference of multiple filters.

What do you think is the feasibility of splitting CPU/CUDA code?
I assume that the extra overhead for CUDA case is coming from the newly added for-loop.

BTW, I found an attempt to use RNN in https://github.com/pytorch/audio/issues/226. Although it is also pointed out that RNN without activation would not improve the performance as much. https://discuss.pytorch.org/t/vanillarnn-without-activation-function/26507/2. I wonder if this could help us in custom CUDA `lfilter` code.